Fix #74: append_exception_reason
