+++ pykern/pkio.py
@@ -8,6 +8,7 @@ from __future__ import absolute_import, division, print_function
 
 # Root module: Limit imports to  avoid dependency issues
 from pykern import pkconst
+from pykern import pkinspect
 import contextlib
 import errno
 import filecmp
@@ -227,7 +228,7 @@ def read_text(filename):
         with open_text(filename) as f:
             return f.read()
     except Exception as e:
-        _exception_reason(e, f"filename={filename}")
+        pkinspect.append_exception_reason(e, f"filename={filename}")
         raise
 
 
@@ -375,25 +376,6 @@ def write_text(path, contents):
         with io.open(str(p), "wt", encoding=TEXT_ENCODING) as f:
             f.write(pkcompat.from_bytes(contents))
     except Exception as e:
-        _exception_reason(e, f"path={path}")
+        pkinspect.append_exception_reason(e, f"path={path}")
         raise
     return p
-
-
-def _exception_reason(exc, reason):
-    def _prefix_reason(string):
-        return ("; " if len(string) > 0 else "") + reason
-
-    if hasattr(exc, "reason") and isinstance(exc.reason, str):
-        exc.reason += _prefix_reason(exc.reason)
-    if hasattr(exc, "args"):
-        if exc.args is None:
-            exc.args = tuple()
-        if isinstance(exc.args, (tuple, list)):
-            if len(exc.args) == 0:
-                exc.args = (reason,)
-            elif isinstance(exc.args[0], str):
-                x = list(exc.args)
-                x[0] += _prefix_reason(x[0])
-                exc.args = tuple(x)
-    # Other cases?
